+++ plugins/upscaledmd/xbrz/xbrz.h
@@ -74,7 +74,7 @@ void nearestNeighborScale(const uint32_t* src, int srcWidth, int srcHeight, int 
                           SliceType st, int yFirst, int yLast);
 
 //parameter tuning
-bool equalColorTest(uint32_t col1, uint32_t col2, ColorFormat colFmt, double luminanceWeight, double equalColorTolerance);
+bool equalColorTest(uint32_t col1, uint32_t col2, ColorFormat colFmt, float luminanceWeight, float equalColorTolerance);
